refactor: tokenLogoLookup
